requirements.txt
django>=4.2
djangorestframework
python-dotenv
requests
PyVCF3